tetragon: Allow to include process' ancestors in exec, exit, kprobe, tracepoint, uprobe and lsm events
Add support for collection of ancestors of the process beyond the immediate parent (up to PID 1 / PID 2) in process_exec, process_exit, process_uprobe, process_kprobe, process_lsm, process_tracepoint events via new configuration options.

If the first attempt to collect all process' ancestors fails, all successfully collected ancestors are added to the event anyway, but the event is added to the eventcache for reprocessing.

In order to guarantee that we are able to collect all ancestor processes, if required options are enabled, 'exec' and 'clone' events now increase reference counters of all ancestors processes of the given process, 'exit' and 'cleanup' events - decrease reference counters of all ancestors processes of the given process. This is true if, and only if, we are able to successfully collect all ancestors of the given process. If we encounter an error during ancestors collection, reference counters of those ancestors will not be changed.

Signed-off-by: t0x01 <T0x01@protonmail.ch>